fix: align machine input and stop behavior with logged transitions
Report pending halt and manual-yield flags ahead of an exhausted input
budget, including when returning a saved rejection tail. Delivering an
accepted input renews the budget, so a yield on its last cycle must not
be mistaken for a sticky overflow. Keep run, log_step, and both collection
APIs consistent without changing interpreter state transitions.

Let send_cmio_response execute the transition's no-ops for inapplicable
boundaries and oversized inputs. Retain API checks for missing, incorrect,
or disallowed revert-root arguments. Use uint32_t response lengths throughout the C and C++ send, log, and
verify APIs, matching the transition. Check Lua and JSON-RPC string
lengths before narrowing so unrepresentable lengths cannot wrap.

Test ordinary and logged no-ops, revert-root validation, oversized uint32_t
response lengths, halt/rejection precedence at overflow, and CLI execution across
a yield on the last budget cycle. Update the API documentation.

+++ src/cm.cpp
@@ -1287,7 +1287,7 @@ cm_error cm_receive_cmio_request(const cm_machine *m, uint8_t *cmd, uint16_t *re
     return cm_result_failure();
 }
 
-cm_error cm_send_cmio_response(cm_machine *m, uint16_t reason, const uint8_t *data, uint64_t length,
+cm_error cm_send_cmio_response(cm_machine *m, uint16_t reason, const uint8_t *data, uint32_t length,
     const cm_hash *revert_root_hash) try {
     auto *cpp_m = convert_from_c(m);
     std::optional<cartesi::machine_hash> cpp_revert_root_hash;
@@ -1300,7 +1300,7 @@ cm_error cm_send_cmio_response(cm_machine *m, uint16_t reason, const uint8_t *da
     return cm_result_failure();
 }
 
-cm_error cm_log_send_cmio_response(cm_machine *m, uint16_t reason, const uint8_t *data, uint64_t length,
+cm_error cm_log_send_cmio_response(cm_machine *m, uint16_t reason, const uint8_t *data, uint32_t length,
     const cm_hash *revert_root_hash, int32_t log_type, const char **log) try {
     if (log == nullptr) {
         throw std::invalid_argument("invalid access log output");
@@ -1319,7 +1319,7 @@ cm_error cm_log_send_cmio_response(cm_machine *m, uint16_t reason, const uint8_t
     return cm_result_failure();
 }
 
-cm_error cm_verify_send_cmio_response(const cm_machine *m, uint16_t reason, const uint8_t *data, uint64_t length,
+cm_error cm_verify_send_cmio_response(const cm_machine *m, uint16_t reason, const uint8_t *data, uint32_t length,
     const cm_hash *root_hash_before, const char *log, const cm_hash *revert_root_hash,
     cm_hash *obtained_root_hash) try {
     if (log == nullptr) {
