+++ src/i-machine.hpp
@@ -308,13 +308,13 @@ public:
     }
 
     /// \brief Sends cmio response.
-    void send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
+    void send_cmio_response(uint16_t reason, const unsigned char *data, uint32_t length,
         std::optional<const_machine_hash_view> revert_root_hash = {}) {
         do_send_cmio_response(reason, data, length, revert_root_hash);
     }
 
     /// \brief Sends cmio response and returns an access log
-    access_log log_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
+    access_log log_send_cmio_response(uint16_t reason, const unsigned char *data, uint32_t length,
         const_machine_hash_view revert_root_hash, const access_log::type &log_type) {
         return do_log_send_cmio_response(reason, data, length, revert_root_hash, log_type);
     }
@@ -351,7 +351,7 @@ public:
     }
 
     /// \brief Checks the validity of state transitions caused by log_send_cmio_response.
-    machine_hash verify_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
+    machine_hash verify_send_cmio_response(uint16_t reason, const unsigned char *data, uint32_t length,
         const_machine_hash_view root_hash_before, const access_log &log,
         const_machine_hash_view revert_root_hash) const {
         return do_verify_send_cmio_response(reason, data, length, root_hash_before, log, revert_root_hash);
@@ -408,9 +408,9 @@ private:
         int32_t log2_bundle_uarch_cycle_count, const machine_hashes &revert_uarch_tail) = 0;
     virtual address_range_descriptions do_get_address_ranges() const = 0;
     virtual machine_cmio_request do_receive_cmio_request(std::span<uint8_t> data) const = 0;
-    virtual void do_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
+    virtual void do_send_cmio_response(uint16_t reason, const unsigned char *data, uint32_t length,
         std::optional<const_machine_hash_view> revert_root_hash) = 0;
-    virtual access_log do_log_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
+    virtual access_log do_log_send_cmio_response(uint16_t reason, const unsigned char *data, uint32_t length,
         const_machine_hash_view revert_root_hash, const access_log::type &log_type) = 0;
     virtual uint64_t do_get_reg_address(reg r) const = 0;
     virtual machine_config do_get_default_config() const = 0;
@@ -421,7 +421,7 @@ private:
         const access_log &log) const = 0;
     virtual machine_hash do_verify_reset_uarch(const_machine_hash_view root_hash_before,
         const access_log &log) const = 0;
-    virtual machine_hash do_verify_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
+    virtual machine_hash do_verify_send_cmio_response(uint16_t reason, const unsigned char *data, uint32_t length,
         const_machine_hash_view root_hash_before, const access_log &log,
         const_machine_hash_view revert_root_hash) const = 0;
     virtual bool do_verify_hash_tree() const = 0;
