+++ src/local-machine.cpp
@@ -230,12 +230,12 @@ machine_cmio_request local_machine::do_receive_cmio_request(std::span<uint8_t> d
     return get_machine()->receive_cmio_request(data);
 }
 
-void local_machine::do_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
+void local_machine::do_send_cmio_response(uint16_t reason, const unsigned char *data, uint32_t length,
     std::optional<const_machine_hash_view> revert_root_hash) {
     get_machine()->send_cmio_response(reason, data, length, revert_root_hash);
 }
 
-access_log local_machine::do_log_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
+access_log local_machine::do_log_send_cmio_response(uint16_t reason, const unsigned char *data, uint32_t length,
     const_machine_hash_view revert_root_hash, const access_log::type &log_type) {
     return get_machine()->log_send_cmio_response(reason, data, length, revert_root_hash, log_type);
 }
@@ -267,7 +267,7 @@ machine_hash local_machine::do_verify_reset_uarch(const_machine_hash_view root_h
     return machine::verify_reset_uarch(root_hash_before, log);
 }
 
-machine_hash local_machine::do_verify_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
+machine_hash local_machine::do_verify_send_cmio_response(uint16_t reason, const unsigned char *data, uint32_t length,
     const_machine_hash_view root_hash_before, const access_log &log, const_machine_hash_view revert_root_hash) const {
     return machine::verify_send_cmio_response(reason, data, length, root_hash_before, log, revert_root_hash);
 }
